+++ tests/test_gwas.py
@@ -7,29 +7,62 @@ Test module for ``gwas.sif`` build
 import os
 import subprocess
 import tempfile
+import platform
 import pytest
 
 
-pth = os.path.join('singularity', 'gwas.sif')
+# Check that (1) singularity exist, and (2) if not, check for docker.
+# If neither are found, tests will fail
+cwd = os.getcwd()
+try:
+    pth = os.path.join(cwd, 'singularity', 'gwas.sif')
+    try:
+        runtime = 'apptainer'
+        subprocess.run(runtime, check=False)
+    except FileNotFoundError:
+        try:
+            runtime = 'singularity'
+            subprocess.run(runtime, check=False)
+        except FileNotFoundError as exc:
+            raise FileNotFoundError from exc
+    PREFIX = f'{runtime} run {pth}'
+    PREFIX_MOUNT = f'{runtime} run --home={cwd}:/home/ ' + '{custom_mount}' + pth
+except FileNotFoundError:
+    try:
+        runtime = 'docker'
+        subprocess.run(runtime, check=False)
+        PREFIX = f'{runtime} run ghcr.io/comorment/gwas'
+        PREFIX_MOUNT = (
+            f'{runtime} run ' +
+            f'--mount type=bind,source={cwd},target={cwd} ' +
+            '{custom_mount}' +
+            '-w /home/ ' +
+            # '--platform linux/amd64 ' +
+            'ghcr.io/comorment/gwas')
+    except FileNotFoundError as err:
+        # neither singularity nor docker found, fall back to plain python
+        # presumably because we are running on the client
+        mssg = 'Neither apptainer, singularity nor docker found, tests will fail'
+        raise FileNotFoundError(mssg) from err
 
 
 def test_gwas_bcftools():
     """test bcftools"""
-    call = f'singularity run {pth} bcftools --version'
+    call = f'{PREFIX} bcftools --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_beagle():
     """test beagle"""
-    call = f'singularity run {pth} beagle'
+    call = f'{PREFIX} beagle'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_bedtools():
     """test bedtools"""
-    call = f'singularity run {pth} bedtools --version'
+    call = f'{PREFIX} bedtools --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
@@ -37,191 +70,234 @@ def test_gwas_bedtools():
 def test_gwas_bgenix():
     """test bgenix, cat-bgen, edit-bgen binaries"""
     for binary in ['bgenix', 'cat-bgen', 'edit-bgen']:
-        call = f'singularity run {pth} {binary} -help'
+        call = f'{PREFIX} {binary} -help'
         out = subprocess.run(call.split(' '), check=False)
         assert out.returncode == 0
 
-
+@pytest.mark.xfail(reason="unsupported on GH builder?", run=True)
 def test_gwas_bolt():
     """test bolt"""
-    call = f'singularity run {pth} bolt -h'
-    out = subprocess.run(call.split(' '), check=False)
-    assert out.returncode == 0
+    call = f'{PREFIX} bolt -h'
+    out = subprocess.run(call, shell=True, check=False, capture_output=True)
+    try:
+        assert out.returncode == 0
+    except AssertionError as exc:
+        print(out.stdout.decode('utf-8'))
+        print(out.stderr.decode('utf-8'))
+        raise AssertionError from exc
 
 
 def test_gwas_eagle():
     """test eagle"""
-    call = f'singularity run {pth} eagle -h'
+    call = f'{PREFIX} eagle -h'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_flashpca():
     """test flashpca"""
-    call = f'singularity run {pth} flashpca_x86-64 --version'
+    call = f'{PREFIX} flashpca_x86-64 --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_gcta():
     """test gcta"""
-    cwd = os.getcwd()
     with tempfile.TemporaryDirectory() as d:
         os.system(f'tar -xvf {cwd}/tests/extras/ex.tar.gz -C {d}')
-        call = f'singularity run --home={d}:/home/ {pth} gcta64 ' + \
-            '--bfile /home/ex --out /home/'
-        out = subprocess.run(call.split(' '), check=False)
+        if runtime == 'docker':
+            custom_mount = f'--mount type=bind,source={d},target={d} '
+        else:
+            custom_mount = f'--bind {d}:{d} '
+        call = f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} gcta64 ' + \
+            f'--bfile {d}/ex --out .'
+        out = subprocess.run(call, shell=True, check=True)
         assert out.returncode == 0
 
 
 def test_gwas_gctb():
     """test gctb"""
-    cwd = os.getcwd()
-    with tempfile.TemporaryDirectory() as d:
-        os.system(f'tar -xvf {cwd}/tests/extras/ex.tar.gz -C {d}')
-        call = f'singularity run {pth} gctb --bfile ex --out .'
-        out = subprocess.run(call.split(' '), check=False)
-        assert out.returncode == 0
+    if platform.machine() == 'arm64':
+        pytest.skip('gctb is not available for arm64 architecture')
+    else:
+        with tempfile.TemporaryDirectory() as d:
+            os.system(f'tar -xvf {cwd}/tests/extras/ex.tar.gz -C {d}')
+            if runtime == 'docker':
+                custom_mount = f'--mount type=bind,source={d},target={d} '
+            else:
+                custom_mount = f'--bind {d}:{d} '
+            call = f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} ' + \
+                f'gctb --bfile {d}/ex --out .'
+            out = subprocess.run(call, shell=True, check=True)
+            assert out.returncode == 0
 
 
 def test_gwas_gwama():
     """test gwama"""
-    call = f'singularity run --home=/tmp/:/home/ {pth} GWAMA --version'
+    call = f'{PREFIX} GWAMA --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_king():
     """test king"""
-    cwd = os.getcwd()
     with tempfile.TemporaryDirectory() as d:
         os.system(f'tar -xvf {cwd}/tests/extras/ex.tar.gz -C {d}')
+        if runtime == 'docker':
+            custom_mount = f'--mount type=bind,source={d},target={d} '
+        else:
+            custom_mount = f'--bind {d}:{d} '
         call = ' '.join(
-            [f'singularity run --home={d}:/home/ {pth} king -b',
-             '/home/ex.bed --fam /home/ex.fam --bim /home/ex.bim --related'])
+            [f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} king -b',
+             f'{d}/ex.bed --fam {d}/ex.fam --bim {d}/ex.bim --related'])
         out = subprocess.run(call.split(' '), check=False)
         assert out.returncode == 0
 
 
 def test_gwas_ldak():
     """test ldak"""
-    call = (f'singularity run {pth} ldak ' +
-            '--make-snps 1 --num-samples 1 --num-snps 1')
-    out = subprocess.run(call.split(' '), check=False)
-    assert out.returncode == 0
+    with tempfile.TemporaryDirectory() as d:
+        if runtime == 'docker':
+            custom_mount = f'--mount type=bind,source={d},target={d} '
+        else:
+            custom_mount = f'--bind {d}:{d} '
+        call = (f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} ldak ' +
+                f'--make-snps {d}/ldak --num-samples 1 --num-snps 1')
+        out = subprocess.run(call, shell=True, check=False)
+        assert out.returncode == 0
 
 
 def test_gwas_liftover():
     """test liftOver"""
-    call = f'singularity run {pth} liftOver -errorHelp'
+    call = f'{PREFIX} liftOver -errorHelp'
     out = subprocess.run(call.split(' '), capture_output=True, check=False)
     assert out.stdout.decode('utf-8').lower().rfind('error') <= 0
 
 
 def test_gwas_metal():
     """test metal"""
-    cwd = os.getcwd()
     with tempfile.TemporaryDirectory() as d:
         os.system(
             f'tar -xvf {cwd}/tests/extras/GlucoseExample.tar.gz -C {d} ' +
             '--strip-components=1')
         os.chdir(d)  # test must be run in temporary directory
+        if runtime == 'docker':
+            custom_mount = f'--mount type=bind,source={d},target={d} '
+        else:
+            custom_mount = f'--bind {d}:{d} '
         call = \
-            f'singularity run {cwd}/{pth} metal metal.txt'
-        out = subprocess.run(call.split(' '), capture_output=True, check=False)
+            f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} metal metal.txt'
+
+        out = subprocess.run(call.replace('-w /home', f'-w {d}'),
+                             shell=True, capture_output=True, check=False)
         assert out.returncode == 0
         # software may not crash on error, checking captured output
         assert out.stdout.decode('utf-8').rfind('Error') <= 0
         assert out.stdout.decode(
             'utf-8').rfind(
                 "## Smallest p-value is 1.491e-12 at marker 'rs560887'") > 0
-    os.chdir(cwd)
 
 
 def test_gwas_minimac4():
     """test minimac4"""
-    call = f'singularity run {pth} minimac4 --version'
+    call = f'{PREFIX} minimac4 --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_plink():
     """test plink"""
-    call = f'singularity run {pth} plink --version'
+    call = f'{PREFIX} plink --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_plink2():
     """test plink2"""
-    call = f'singularity run {pth} plink2 --version'
+    call = f'{PREFIX} plink2 --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_plink2_avx2():
     """test plink2_avx2"""
-    call = f'singularity run {pth} plink2_avx2 --version'
-    out = subprocess.run(call.split(' '), check=False)
-    assert out.returncode == 0
+    if platform.machine() == 'arm64':
+        pytest.skip('plink2_avx2 is not available for arm64 architecture')
+    else:
+        call = f'{PREFIX} plink2_avx2 --version'
+        out = subprocess.run(call.split(' '), check=False)
+        assert out.returncode == 0
 
 
 def test_gwas_prsice():
     """test prsice"""
-    call = f'singularity run {pth} PRSice_linux --version'
+    call = f'{PREFIX} PRSice_linux --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_qctools():
     """test qctools"""
-    for binary in ['inthinnerator', 'hptest', 'ldbird', 'qctool', 'selfmap']:
-        call = f'singularity run {pth} {binary} -help'
-        out = subprocess.run(call.split(' '), check=False)
-        assert out.returncode == 0
+    if platform.machine() == 'arm64':
+        pytest.skip('qctools is not available for arm64 architecture')
+    else:
+        for binary in ['inthinnerator', 'hptest',
+                       'ldbird', 'qctool', 'selfmap']:
+            call = f'{PREFIX} {binary} -help'
+            out = subprocess.run(call.split(' '), check=False)
+            assert out.returncode == 0
 
 
 def test_gwas_regenie():
     """test regenie"""
-    call = f'singularity run {pth} regenie option --help'
+    call = f'{PREFIX} regenie option --help'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_samtools():
     """test samtools"""
-    call = f'singularity run {pth} samtools --help'
+    call = f'{PREFIX} samtools --help'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_shapeit4():
     """test shapeit4"""
-    call = f'singularity run {pth} shapeit4.2 --help'
-    out = subprocess.run(call.split(' '), check=False)
-    assert out.returncode == 0
-
-
-def test_gwas_shapeit5():
-    """test shapeit5"""
-    for binary in ['phase_common', 'phase_rare',
-                   'ligate', 'switch', 'xcftools']:
-        call = f'singularity run {pth} {binary} --help'
+    if platform.machine() == 'arm64':
+        pytest.skip('shapeit4 is not available for arm64 architecture')
+    else:
+        call = f'{PREFIX} shapeit4.2 --help'
         out = subprocess.run(call.split(' '), check=False)
         assert out.returncode == 0
 
 
+def test_gwas_shapeit5():
+    """test shapeit5"""
+    if platform.machine() == 'arm64':
+        pytest.skip('shapeit5 is not available for arm64 architecture')
+    else:
+        for binary in ['phase_common', 'phase_rare',
+                       'ligate', 'switch', 'xcftools']:
+            call = f'{PREFIX} {binary} --help'
+            out = subprocess.run(call.split(' '), check=False)
+            assert out.returncode == 0
+
+
 def test_gwas_snptest():
     """test snptest"""
-    call = f'singularity run {pth} snptest -help'
-    out = subprocess.run(call.split(' '), check=False)
-    assert out.returncode == 0
+    if platform.machine() == 'arm64':
+        pytest.skip('snptest is not available for arm64 architecture')
+    else:
+        call = f'{PREFIX} snptest -help'
+        out = subprocess.run(call.split(' '), check=False)
+        assert out.returncode == 0
 
 
 @pytest.mark.xfail(reason="no help function for switchError")
 def test_gwas_switchError():
     """test switchError"""
-    call = (f'singularity run {pth} switchError' +
+    call = (f'{PREFIX} switchError' +
             '--reg foo --gen foo --hap foo --fam foo ' +
             '--ps foo --out foo --maf 0.0')
     out = subprocess.run(call.split(' '), check=False)
@@ -230,12 +306,15 @@ def test_gwas_switchError():
 
 def test_gwas_simu():
     """test simu"""
-    cwd = os.getcwd()
     with tempfile.TemporaryDirectory() as d:
         os.system(f'tar -xvf {cwd}/tests/extras/ex.tar.gz -C {d}')
+        if runtime == 'docker':
+            custom_mount = f'--mount type=bind,source={d},target={d} '
+        else:
+            custom_mount = f'--bind {d}:{d} '
         call = ' '.join(
-            [f'singularity run --home={d}:/home/ {pth}',
-             'simu_linux --bfile /home/ex --qt ',
+            [f'{PREFIX_MOUNT.format(custom_mount=custom_mount)} ' +
+             f'simu_linux --bfile {d}/ex --qt ',
              '--causal-pi 0.01 --num-traits 2 --hsq 0.2 0.6 --rg 0.8'])
         out = subprocess.run(call.split(' '), check=False)
         assert out.returncode == 0
@@ -243,13 +322,13 @@ def test_gwas_simu():
 
 def test_gwas_vcftools():
     """test vcftools"""
-    call = f'singularity run {pth} vcftools --version'
+    call = f'{PREFIX} vcftools --version'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 
 def test_gwas_duohmm():
     """test duohmm"""
-    call = f'singularity run {pth} duohmm'
+    call = f'{PREFIX} duohmm'
     out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0