+++ tests/test_python3.py
@@ -5,51 +5,83 @@ Test module for ``python3.sif`` build
 """
 
 import os
+import socket
 import subprocess
 
-pth = os.path.join('singularity', 'python3.sif')
+# port used by tests
+sock = socket.socket()
+sock.bind(('', 0))
+port = sock.getsockname()[1]
 
+# Check that (1) singularity exist, and (2) if not, check for docker.
+# If neither are found, tests will fail
+cwd = os.getcwd()
+try:
+    pth = os.path.join('singularity', 'python3.sif')
+    subprocess.run('singularity', check=False)
+    PREFIX = f'singularity run {pth}'
+    PREFIX_MOUNT = f'singularity run --home={cwd}:/home/ {pth}'
+    PYTHON = f'{PREFIX} python'
+    PYTHON_MOUNT = f'{PREFIX_MOUNT} python'
+    PLINK = f'{PREFIX} plink'
+    PRSICE = f'{PREFIX} PRSice_linux'
+except FileNotFoundError:
+    try:
+        subprocess.run('docker', check=False)
+        PREFIX = (f'docker run -p {port}:{port} ' +
+                  'ghcr.io/comorment/python3')
+        PREFIX_MOUNT = (
+            f'docker run -p {port}:{port} ' +
+            f'--mount type=bind,source={cwd},target={cwd} ' +
+            '--platform linux/amd64 ' +
+            'ghcr.io/comorment/python3')
+        PYTHON = f'{PREFIX} python'
+        PYTHON_MOUNT = f'{PREFIX_MOUNT} python'
+        PLINK = f'{PREFIX} plink'
+        PRSICE = f'{PREFIX} PRSice_linux'
+    except FileNotFoundError:
+        # neither singularity nor docker found, fall back to plain python
+        # presumably because we are running on the client
+        PYTHON = 'python'
+        PYTHON_MOUNT = 'python'
+        PLINK = 'plink'
+        PRSICE = 'PRSice_linux'
 
 def test_python3_plink():
-    call = f'singularity run {pth} plink --version'
-    out = subprocess.run(call.split(' '))
+    call = f'{PLINK} --version'
+    out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 def test_python3_prsice():
-    call = f'singularity run {pth} PRSice_linux --version'
-    out = subprocess.run(call.split(' '))
+    call = f'{PRSICE} --version'
+    out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
 
 def test_python3_python():
-    call = f'singularity run {pth} python --version'
-    out = subprocess.run(call.split(' '))
+    call = f'{PYTHON} --version'
+    out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
-
 
 def test_python3_fastlmm():
-    pwd = os.getcwd()
-    call = f'singularity run --home={pwd} {pth} python ' + \
-           f'{pwd}/tests/extras/fastlmm_example_script.py'
-    out = subprocess.run(call.split(' '))
+    call = f'{PYTHON_MOUNT} ' + \
+          f'{cwd}/tests/extras/fastlmm_example_script.py'
+    out = subprocess.run(call, shell=True, check=False)
     assert out.returncode == 0
-
 
 def test_python3_python_convert():
     args = ['/tools/python_convert/sumstats.py -h',
             '/tools/python_convert/tests/test_consistent.py',
             '/tools/python_convert/tests/test_duplicated.py']
     for arg in args:
-        call = f'singularity run {pth} python {arg}'
-        out = subprocess.run(call.split(' '))
+        call = f'{PYTHON} {arg}'
+        out = subprocess.run(call.split(' '), check=False)
         assert out.returncode == 0
-
 
 def test_python3_ukb():
     arg = '/tools/ukb/ukb_helper.py -h'
-    call = f'singularity run {pth} python {arg}'
-    out = subprocess.run(call.split(' '))
+    call = f'{PYTHON} {arg}'
+    out = subprocess.run(call.split(' '), check=False)
     assert out.returncode == 0
-
 
 def test_python3_packages():
     packages = [
@@ -82,14 +114,12 @@ def test_python3_packages():
         'xmltodict',
         'yaml',  # pyyaml
         ]
-    for pkg in packages:
-        call = f'singularity run {pth} python -c "import {pkg}"'
-        out = subprocess.run(call.split(' '))
-        assert out.returncode == 0
-
+    importstr = 'import ' + ', '.join(packages)
+    call = f"{PYTHON} -c '{importstr}'"
+    out = subprocess.run(call, shell=True, check=False)
+    assert out.returncode == 0
 
 def test_python3_import_pandas_scipy_stats():
-    pwd = os.getcwd()
-    call = f'singularity run --home={pwd} {pth} python -c "import pandas as pd; from scipy import *"'
-    out = subprocess.run(call.split(' '))
+    call = f'{PYTHON_MOUNT} -c "import pandas as pd; from scipy import *"'
+    out = subprocess.run(call, shell=True, check=False)
     assert out.returncode == 0