+++ CHANGELOG.md
@@ -21,6 +21,7 @@ If MD5 sum is not listed for a certain release then it means that the container 
 
 ### Added
 
+* Add unit test runs as part of the GitHub Actions workflow for building Docker containers
 * Add Python packages `imbalanced-learn, lightgbm, openpyxl` + PRSice_linux binary to `python3.sif` container
 * Add Conda environment file for project dependencies
 * Add Python packages `scikit-survival, pandas-plink, numba, xmltodict, pyliftover, configparser, intervaltree` to `python3.sif` container
@@ -36,6 +37,8 @@ If MD5 sum is not listed for a certain release then it means that the container 
 
 ### Updated
 
+* update testing scripts to support both Docker and Singularity containers
+* Update REGENIE binary to version 3.6 in `gwas.sif` container
 * Update LDAK binary to version 6 in gwas.sif (from 5.2)
 * Rebuilt `gwas.sif` container with md5sum checksum: